* src/util.c (encode_stored_crc): use size_t.
* src/lha.h (struct LzHeader): use size_t for the members
`packed_size' and `original_size'.


git-svn-id: svn+ssh://svn.sourceforge.jp/svnroot/lha/lha/trunk@708 6a8cc165-1e22-0410-a132-eb4e3f353aba

+++ src/lha.h
@@ -241,8 +241,8 @@ typedef struct LzHeader {
     size_t          header_size;
     int             size_field_length;
     char            method[METHOD_TYPE_STORAGE];
-    long            packed_size;
-    long            original_size;
+    size_t          packed_size;
+    size_t          original_size;
     unsigned char   attribute;
     unsigned char   header_level;
     char            name[FILENAME_LENGTH];
